Moved all code into one file
