+++ .gitignore
@@ -55,3 +55,6 @@ docs/_build/
 
 # PyBuilder
 target/
+/.idea
+/*.pybak
+/*.pylog
