--- 3_tone_splicing.pylog
@@ -1,27 +1,0 @@
-<file_created time="Sat Jan 09 11:57:22 GMT 2016">
-<Could not find G:\Documents\GitHub\COMP120\comp-120-tinkering-audio\3_tone_splicing.pylog, creating a new logfile>
-<menu_button time="Sat Jan 09 11:57:31 GMT 2016">Load Program</menu_button>
-<command_run time="Sat Jan 09 11:58:01 GMT 2016">t1 = toneGen(44, 2000,1)></command_run>
-<menu_button time="Sat Jan 09 11:58:04 GMT 2016">Load Program</menu_button>
-<file_saved time="Sat Jan 09 11:58:05 GMT 2016">G:\Documents\GitHub\COMP120\comp-120-tinkering-audio\3_tone_splicing.py</file_saved>
-</file_created>
-<command_run time="Sat Jan 09 11:58:10 GMT 2016">t1 = toneGen(44, 2000,1)></command_run>
-<command_run time="Sat Jan 09 11:58:13 GMT 2016">play(t1)></command_run>
-<command_run time="Sat Jan 09 11:58:26 GMT 2016">t1 = toneGen(440, 2000,1)></command_run>
-<command_run time="Sat Jan 09 11:58:27 GMT 2016">play(t1)></command_run>
-<command_run time="Sat Jan 09 11:58:40 GMT 2016">t2 = toneGen(880, 2000,1)></command_run>
-<command_run time="Sat Jan 09 11:58:43 GMT 2016">play(t2)></command_run>
-<command_run time="Sat Jan 09 11:59:25 GMT 2016">audioSplice(t1, t2, 2)></command_run>
-<command_run time="Sat Jan 09 11:59:27 GMT 2016">audioSplice(t1, t2, 3)></command_run>
-<command_run time="Sat Jan 09 12:13:17 GMT 2016">spliceAndSwap()></command_run>
-<menu_button time="Sat Jan 09 12:13:19 GMT 2016">Load Program</menu_button>
-<file_saved time="Sat Jan 09 12:13:20 GMT 2016">G:\Documents\GitHub\COMP120\comp-120-tinkering-audio\3_tone_splicing.py</file_saved>
-</file_created>
-<command_run time="Sat Jan 09 12:13:21 GMT 2016">spliceAndSwap()></command_run>
-<menu_button time="Sat Jan 09 12:13:31 GMT 2016">Load Program</menu_button>
-<file_saved time="Sat Jan 09 12:13:31 GMT 2016">G:\Documents\GitHub\COMP120\comp-120-tinkering-audio\3_tone_splicing.py</file_saved>
-</file_created>
-<command_run time="Sat Jan 09 12:13:32 GMT 2016">spliceAndSwap()></command_run>
-<menu_button time="Sat Jan 09 12:14:22 GMT 2016">Load Program</menu_button>
-<file_saved time="Sat Jan 09 12:14:23 GMT 2016">G:\Documents\GitHub\COMP120\comp-120-tinkering-audio\3_tone_splicing.py</file_saved>
-</file_created>
